building team functionality

+++ application/views/individual.php
@@ -20,18 +20,24 @@
 	
 	<ol class="taskList">
 	<?
-	foreach ($results as $row) {
-		?>
-		<li class="taskItem">
-			<? echo $row->fldName; ?>
-			
-		<div class="taskOptions">
-			<a href='/tasks/view/<? echo uri_string()."/".$row->pkTaskId; ?>' rel="#overlay" title="Information" >i</a>
-			<a href='/tasks/assignTo/<? echo uri_string()."/".$row->pkTaskId; ?>' title="Assign Task">+</a>
-			<a href='/tasks/delete/<? echo uri_string()."/".$row->pkTaskId; ?>' title="Delete" class="deleteTask">X</a>
-		</div>
-		</li>
-		<?
+	if(count($results)==0)
+	{
+		echo '<li class="taskItem">No tasks exist for user <span class="showUsername">'.$user.'</span> OR user <span class="showUsername">'.$user.'</span> doesn'."'".'t exist</li>';
+	}
+	else{
+		foreach ($results as $row) {
+			?>
+			<li class="taskItem">
+				<? echo $row->fldName; ?>
+				
+			<div class="taskOptions">
+				<a href='/tasks/view/<? echo uri_string()."/".$row->pkTaskId; ?>' rel="#overlay" title="Information" >i</a>
+				<a href='/tasks/assignTo/<? echo uri_string()."/".$row->pkTaskId; ?>' title="Assign Task">+</a>
+				<a href='/tasks/delete/<? echo uri_string()."/".$row->pkTaskId; ?>' title="Delete" class="deleteTask">X</a>
+			</div>
+			</li>
+			<?
+		}
 	}
 	?>
 	</ol>
